LUCENE-2111: fix standard codec's terms dict to handle > 2B terms
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/branches/flex_1458@929908 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/lucene/index/codecs/standard/SimpleStandardTermsIndexReader.java
@@ -60,7 +60,15 @@ import org.apache.lucene.index.IndexFileNames;
 /** @lucene.experimental */
 public class SimpleStandardTermsIndexReader extends StandardTermsIndexReader {
 
-  private int totalIndexInterval;
+  // NOTE: long is overkill here, since this number is 128
+  // by default and only indexDivisor * 128 if you change
+  // the indexDivisor at search time.  But, we use this in a
+  // number of places to multiply out the actual ord, and we
+  // will overflow int during those multiplies.  So to avoid
+  // having to upgrade each multiple to long in multiple
+  // places (error proned), we use long here:
+  private long totalIndexInterval;
+
   private int indexDivisor;
   final private int indexInterval;
